Use appropriate updates proxy if it is qrexec based
In Qubes 4.0, updates proxy runs on qrexec and templates have no other
network access. This is accompanied with local TCP listener redirecting
the traffic to the proxy.
If the listener is active, use its address.
This variable is (currently) used only to check if the proxy is
torified. Actual proxy setting in apt is done by standard Qubes startup
scripts.